==> src/visualisation/0-residential-segregation-vg.ipynb <==
# Data exploration

%load_ext autoreload
%autoreload 2
%cd D:\mobi-social-segregation-se

import pandas as pd
import sqlalchemy
import preprocess
import numpy as np
import matplotlib.pyplot as plt
import geopandas as gpd

user = preprocess.keys_manager['database']['user']
password = preprocess.keys_manager['database']['password']
port = preprocess.keys_manager['database']['port']
db_name = preprocess.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}')

## 1 Residential segregation

gdf = gpd.GeoDataFrame.from_postgis("SELECT deso, geom FROM public.zones", con=engine)
df_socio = pd.read_sql_query(sql="""SELECT * FROM network.selected_socioeco_factors;""", con=engine)
socio_factors = ['deso', 'Education - pre-secondary','Education - post-secondary education',
                 'Population - foreign background', 'Population - Swedish',
                 'Population - European', 'Population - The rest of the world',
                 'Population - Living - tenancy', 'Non-employed',
                 'Net income - median KSEK', 'Low-economic standard percent']
df_socio = df_socio.loc[:, socio_factors].set_index('deso').reindex(gdf.deso.tolist()).sort_index().reset_index()

df_socio.iloc[0]

df_socio = df_socio.loc[:, ['deso', 'Population - Swedish', 'Net income - median KSEK']].rename(columns={'Population - Swedish': 'pop_se','Net income - median KSEK': 'income'})
df_socio.loc[:, 'pop_nse'] = 1 - df_socio.loc[:, 'pop_se']

df_socio.loc[:, ['deso', 'pop_nse', 'income']].to_csv('results/demo.csv', index=False)
